backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
sentencepiece==0.1.99
protobuf>=3.19.6,<4.24
pydantic==2.5.3
python-multipart==0.0.6
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Voice dependencies
openai-whisper==20231117
sounddevice==0.4.6
soundfile==0.12.1
pydub==0.25.1
pyaudio==0.2.14
gtts==2.5.0
pyttsx3==2.90

# Configuration dependencies
PyYAML==6.0.1
edge-tts==6.1.9
websockets==12.0
numpy==1.24.3
scipy==1.11.4

# Audio processing dependencies
noisereduce==3.0.0
librosa==0.10.1
numba==0.58.1  # Required by librosa
audioread==3.0.1
resampy==0.4.2

# Automation dependencies
icalendar==5.0.11
pytz==2023.3
apscheduler==3.10.4
aiohttp==3.9.1
requests==2.31.0
httpx==0.27.0

# RAG dependencies
sentence-transformers==2.3.0
faiss-cpu==1.7.4
chromadb==0.4.22
tiktoken==0.5.2
scikit-learn==1.3.2
nltk==3.8.1
aiofiles==23.2.1

# Model training dependencies
wandb==0.16.2
tensorboard==2.15.1
datasets==2.16.1
evaluate==0.4.1
rouge-score==0.1.2

# Memory management dependencies
psutil==5.9.8
objgraph==3.6.1
pympler==1.0.1

# LangChain and Tools
langchain==0.0.350
langchain-community==0.0.10
langchain-experimental==0.0.47
duckduckgo-search==4.1.1
numexpr==2.8.8  # For calculator tool
wikipedia-api==0.6.0
python-dotenv==1.0.0

# Cloud AI APIs
anthropic==0.18.1
huggingface-hub==0.19.4  # For Claude API integration

# Weather and location services
geocoder>=1.38.0  # For IP-based location detection

# ML Enhancement dependencies
joblib>=1.3.0  # For model persistence
torchaudio==2.1.2  # For deep audio processing (must match torch version)
SpeechRecognition>=3.10.0  # For enhanced speech recognition
pygame>=2.5.0  # For audio feedback
# Note: scikit-learn is already included above as version 1.3.2

# Computer Vision dependencies
opencv-python>=4.8.0  # For image processing
pytesseract>=0.3.10  # For OCR text extraction
Pillow>=10.0.0  # For image manipulation
pyobjc-core==10.1  # Core PyObjC for macOS - specific version for miniforge compatibility
pyobjc-framework-Cocoa==10.1  # Required base framework
pyobjc-framework-Quartz==10.1  # For macOS screen capture (includes CoreGraphics)
pyobjc-framework-Vision==10.1  # For macOS Vision framework
pyobjc-framework-CoreML==10.1  # Required by Vision framework
# Note: PyAutoGUI removed due to compatibility issues with miniforge - using native macOS APIs instead